Add tests for Data, Date, URL
